exclude EntryDataBase for from EntryDataBase.h
